# adapter/src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: landmark area API
  description: landmark service
  version: "1.0"
paths:
...
  /administrative-areas:
    post:
      summary: Create area
      operationId: createAdministrativeArea
      tags:
        - administrativeArea
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                request:
                  $ref: "#/components/schemas/AdministrativeAreasImportRequest"
                geometry:
                  type: object
      responses:
        "400":
          description: Bad Request
        "201":
          description: Created
  /administrative-areas/{AdministrativeAreaId}:
    get:
      summary: display single administrative area
      operationId: loadAdministrativeArea
      tags:
        - administrativeArea
      parameters:
        - in: path
          name: AdministrativeAreaId
          required: true
          schema:
            type: AdministrativeAreaId
      responses:
        "400":
          description: Bad Request
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdministrativeAreaResponse"
    put:
      summary: update single administrative area
      operationId: updateAdministrativeArea
      tags:
        - administrativeArea
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdministrativeAreaRequest"
        required: true
      parameters:
        - in: path
          name: AdministrativeAreaId
          required: true
          schema:
            type: AdministrativeAreaId
      responses:
        "400":
          description: Bad Request
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdministrativeAreaResponse"
    delete:
      summary: delete single administrative area
      operationId: deleteAdministrativeArea
      tags:
        - administrativeArea
      parameters:
        - in: path
          name: AdministrativeAreaId
          required: true
          schema:
            type: AdministrativeAreaId
      responses:
        "400":
          description: Bad Request
        "204":
          description:
  /administrative-areas/by-level-country:
    get:
      summary: Get administrative areas by level and country name
      operationId: getAdministrativeAreasByLevelAndCountry
      tags:
        - administrativeArea
      parameters:
        - in: query
          name: levelName
          required: true
          schema:
            type: string
        - in: query
          name: countryName
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of administrative areas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdministrativeAreaResponse"
  /administrative-areas/by-metadata:
    get:
      summary: Get administrative areas by metadata query
      operationId: getAdministrativeAreasByMetadata
      tags:
        - administrativeArea
      parameters:
        - in: query
          name: metadataQuery
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of administrative areas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdministrativeAreaResponse"
...
components:
  schemas:
    AdministrativeAreasImportRequest:
      type: object
      required:
        - level
      properties:
        level:
          $ref: "#/components/schemas/AdministrativeLevelBasic"
        metadata:
          $ref: "#/components/schemas/LandmarkMetadata"
    AdministrativeAreaRequest:
      type: object
      required:
        - name
        - level
      properties:
        name:
          type: LandmarkName
        level:
          $ref: "#/components/schemas/AdministrativeLevelBasic"
        geometry:
          type: string
        metadata:
          $ref: "#/components/schemas/LandmarkMetadata"
    AdministrativeAreaResponse:
      type: object
      required:
        - name
      properties:
        id:
          type: AdministrativeAreaId
        name:
          type: LandmarkName
        createTime:
          type: string
          format: date-time
        updateTime:
          type: string
          format: date-time
        level:
          $ref: "#/components/schemas/AdministrativeLevel"
        parent:
          type: AdministrativeAreaId
        geometry:
          type: string
        metadata:
          $ref: "#/components/schemas/LandmarkMetadata"
    AdministrativeLevelBasic:
      type: object
      required:
        - country
        - name
      properties:
        name:
          type: LevelName
        country:
          type: CountryName
    AdministrativeLevel:
      type: object
      required:
        - order
        - country
        - name
      properties:
        id:
          type: AdministrativeLevelId
        name:
          type: LevelName
        createTime:
          type: string
          format: date-time
        updateTime:
          type: string
          format: date-time
        country:
          type: CountryName
        order:
          type: LevelOrder
...
    LandmarkMetadata:
      type: object